requirements.txt
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
cohere>=4.37
huggingface-hub>=0.19.0
python-dotenv>=1.0.0
requests>=2.31.0
colorama>=0.4.6
